requirements.txt
pandas~=2.2.2
tqdm~=4.66.5
requests~=2.32.3
numpy~=2.1.1
pillow~=10.4.0
plotly~=5.24.1
emoji~=2.13.2
certifi~=2024.8.30
geopy~=2.4.1
mercantile~=1.2.1
reverse-geocode~=1.6.5
plotly~=5.24.1
mlflow~=2.17.0rc0
python-dotenv~=1.0.1
ipywidgets~=8.1.5